app_python/pyproject.toml
[project]
name = "app-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "redis>=4.2.0,<5.0.0",
    "fastapi-cache2[redis]>=0.2.1"
]
